pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "endgame-fen-generator"
version = "0.1.0"
description = "Extract realistic chess training positions from large Lichess PGN archives."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-chess>=1.999",
  "PyYAML>=6.0.1",
  "zstandard>=0.22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
extract-training-positions = "training_positions.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
